+++ mcp_tools/container_scan/clair_vulnerability.py
@@ -1,0 +1,34 @@
+# mcp_tools/container_scan/clair_vulnerability.py
+
+from typing import Dict, Any
+
+def register_clair_vulnerability_tool(mcp, hexstrike_client, logger):
+    
+    @mcp.tool()
+    def clair_vulnerability_scan(image: str, config: str = "/etc/clair/config.yaml",
+                                output_format: str = "json", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Clair for container vulnerability analysis.
+
+        Args:
+            image: Container image to scan
+            config: Clair configuration file
+            output_format: Output format (json, yaml)
+            additional_args: Additional Clair arguments
+
+        Returns:
+            Container vulnerability analysis results
+        """
+        data = {
+            "image": image,
+            "config": config,
+            "output_format": output_format,
+            "additional_args": additional_args
+        }
+        logger.info(f"🐳 Starting Clair vulnerability scan: {image}")
+        result = hexstrike_client.safe_post("api/tools/clair", data)
+        if result.get("success"):
+            logger.info(f"✅ Clair scan completed for {image}")
+        else:
+            logger.error(f"❌ Clair scan failed for {image}")
+        return result
